+++ src/main/java/codesquad/Main.java
@@ -11,10 +11,10 @@ import codesquad.application.model.business.LoginUserLogic;
 import codesquad.application.model.business.RegisterUserLogic;
 import codesquad.application.model.business.user.GetUserInfoLogic;
 import codesquad.application.model.business.user.GetUserListLogic;
-import codesquad.application.post.argumentresolver.PostCreateArgumentResolver;
-import codesquad.application.post.business.PostCreateLogic;
-import codesquad.application.post.handler.PostCreateRequestHandler;
-import codesquad.application.post.request.PostCreateRequest;
+import codesquad.application.domain.post.argumentresolver.PostCreateArgumentResolver;
+import codesquad.application.domain.post.business.PostCreateLogic;
+import codesquad.application.domain.post.handler.PostCreateRequestHandler;
+import codesquad.application.domain.post.request.PostCreateRequest;
 import codesquad.application.processor.HandlerRegistry;
 import codesquad.application.processor.HttpRequestDispatcher;
 import codesquad.application.processor.argumentresolver.ArgumentResolver;
