+++ src/main/java/codesquad/application/domain/post/argumentresolver/PostCreateArgumentResolver.java
@@ -1,7 +1,7 @@
-package codesquad.application.post.argumentresolver;
+package codesquad.application.domain.post.argumentresolver;
 
 import codesquad.api.Request;
-import codesquad.application.post.request.PostCreateRequest;
+import codesquad.application.domain.post.request.PostCreateRequest;
 import codesquad.application.processor.argumentresolver.ArgumentResolver;
 import codesquad.webserver.helper.MultiPartParseHelper;
 import codesquad.webserver.http.header.HttpHeaders;
