--- src/main/java/codesquad/application/web/post/PostRequest.java
@@ -1,10 +1,0 @@
-package codesquad.application.web.post;
-
-import java.io.ByteArrayInputStream;
-
-public record PostRequest(
-        Long userId,
-        String content,
-        ByteArrayInputStream image
-) {
-}
